requirements.txt
python-kasa[speedups]
